backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
ecdsa==0.18.0
fastapi==0.104.1
h11==0.14.0
idna==3.6
jwt==1.3.1
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
